tests/wpt/tests/html/canvas/element/path-objects/2d.path.arcTo.ensuresubpath.2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.arcTo.ensuresubpath.2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.arcTo.ensuresubpath.2</h1>
<p class="desc">If there is no subpath, the first control point is added</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("If there is no subpath, the first control point is added");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.lineWidth = 50;
  ctx.strokeStyle = '#0f0';
  ctx.beginPath();
  ctx.arcTo(0, 25, 50, 250, 0.1); // adds (x1,y1), draws nothing
  ctx.lineTo(100, 25);
  ctx.stroke();
  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
